requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
mysqlclient>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.2.0
pyyaml>=6.0

# Web automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Social media APIs
tweepy>=4.14.0
requests-oauthlib>=1.3.1

# LLM APIs - Gemini
google-generativeai>=0.3.0

# Image processing
Pillow>=10.0.0

# Scheduling
apscheduler>=3.10.0

# Database
alembic>=1.12.0

# Utilities
python-dateutil>=2.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
